Organized all OpModes into "Auto", "TeleOp", or "Concept" packages - nothing was deleted!
