--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/opmode/FollowerPIDTuner.java
@@ -1,39 +1,0 @@
-package org.firstinspires.ftc.teamcode.drive.opmode;
-
-import com.acmerobotics.dashboard.config.Config;
-import com.acmerobotics.roadrunner.geometry.Pose2d;
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-
-import org.firstinspires.ftc.teamcode.drive.mecanum.SampleMecanumDriveBase;
-import org.firstinspires.ftc.teamcode.drive.mecanum.SampleMecanumDriveREV;
-
-/*
- * Op mode for tuning follower PID coefficients (located in the drive base classes). The robot
- * drives in a DISTANCE-by-DISTANCE square indefinitely.
- */
-@Config
-@Autonomous(group = "drive")
-public class FollowerPIDTuner extends LinearOpMode {
-    public static double DISTANCE = 48;
-
-    @Override
-    public void runOpMode() throws InterruptedException {
-        SampleMecanumDriveBase drive = new SampleMecanumDriveREV(hardwareMap);
-
-        drive.setPoseEstimate(new Pose2d(-DISTANCE / 2, -DISTANCE / 2, 0));
-
-        waitForStart();
-
-        if (isStopRequested()) return;
-
-        while (!isStopRequested()) {
-            drive.followTrajectorySync(
-                    drive.trajectoryBuilder()
-                            .forward(DISTANCE)
-                            .build()
-            );
-            drive.turnSync(Math.toRadians(90));
-        }
-    }
-}
