+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/roadrunner/opmode/SplineTest.java
@@ -1,0 +1,38 @@
+package org.firstinspires.ftc.teamcode.roadrunner.opmode;
+
+import com.acmerobotics.roadrunner.geometry.Pose2d;
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+
+import org.firstinspires.ftc.teamcode.roadrunner.mecanum.SampleMecanumDriveBase;
+import org.firstinspires.ftc.teamcode.roadrunner.mecanum.SampleMecanumDriveREV;
+
+/*
+ * This is an example of a more complex path to really test the tuning.
+ */
+@Autonomous(group = "drive")
+public class SplineTest extends LinearOpMode {
+    @Override
+    public void runOpMode() throws InterruptedException {
+        SampleMecanumDriveBase drive = new SampleMecanumDriveREV(hardwareMap);
+
+        waitForStart();
+
+        if (isStopRequested()) return;
+
+        drive.followTrajectorySync(
+                drive.trajectoryBuilder()
+                        .splineTo(new Pose2d(30, 30, 0))
+                        .build()
+        );
+
+        sleep(2000);
+
+        drive.followTrajectorySync(
+                drive.trajectoryBuilder()
+                        .reverse()
+                        .splineTo(new Pose2d(0, 0, 0))
+                        .build()
+        );
+    }
+}
