+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/roadrunner/opmode/StraightTest.java
@@ -1,0 +1,33 @@
+package org.firstinspires.ftc.teamcode.roadrunner.opmode;
+
+import com.acmerobotics.dashboard.config.Config;
+import com.acmerobotics.roadrunner.trajectory.Trajectory;
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+
+import org.firstinspires.ftc.teamcode.roadrunner.mecanum.SampleMecanumDriveBase;
+import org.firstinspires.ftc.teamcode.roadrunner.mecanum.SampleMecanumDriveREV;
+
+/*
+ * This is a simple routine to test translational drive capabilities.
+ */
+@Config
+@Autonomous(group = "drive")
+public class StraightTest extends LinearOpMode {
+    public static double DISTANCE = 60;
+
+    @Override
+    public void runOpMode() throws InterruptedException {
+        SampleMecanumDriveBase drive = new SampleMecanumDriveREV(hardwareMap);
+
+        Trajectory trajectory = drive.trajectoryBuilder()
+                .forward(DISTANCE)
+                .build();
+
+        waitForStart();
+
+        if (isStopRequested()) return;
+
+        drive.followTrajectorySync(trajectory);
+    }
+}
