+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/roadrunner/opmode/TurnTest.java
@@ -1,0 +1,28 @@
+package org.firstinspires.ftc.teamcode.roadrunner.opmode;
+
+import com.acmerobotics.dashboard.config.Config;
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+
+import org.firstinspires.ftc.teamcode.roadrunner.mecanum.SampleMecanumDriveBase;
+import org.firstinspires.ftc.teamcode.roadrunner.mecanum.SampleMecanumDriveREV;
+
+/*
+ * This is a simple routine to test turning capabilities.
+ */
+@Config
+@Autonomous(group = "drive")
+public class TurnTest extends LinearOpMode {
+    public static double ANGLE = 90; // deg
+
+    @Override
+    public void runOpMode() throws InterruptedException {
+        SampleMecanumDriveBase drive = new SampleMecanumDriveREV(hardwareMap);
+
+        waitForStart();
+
+        if (isStopRequested()) return;
+
+        drive.turnSync(Math.toRadians(ANGLE));
+    }
+}
